% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_lag.R
\name{add_time_lag}
\alias{add_time_lag}
\title{Add Time Lags Between Sampling}
\usage{
add_time_lag(ps, time.col = "Time_hr_num", sam.name.col = "FAIR_Labels")
}
\arguments{
\item{ps}{A \code{\link{phyloseq-class}}.}

\item{time.col}{Name of column in sample_data i.e. metadata specifying
sampling time in hours or days or mins}

\item{sam.name.col}{Name of column in sample_data i.e. metadata specifying
sample names. This should match the output of sample_names(ps).}
}
\value{
A \code{\link{phyloseq-class}} with addition of time span between two sampling
        events to the metadata.
}
\description{
For longitudinal sampling, intervals between
sampling are identified and added to the sample_data.
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
ps1.b5.lag <- add_time_lag(ps1.b5, time.col = "Time_hr_num", sam.name.col = "FAIR_Labels")
head(meta(ps1.b5.lag))
}

}
\references{
See citation("syncomR")
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Utilities}
