% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{syncom_colors}
\alias{syncom_colors}
\alias{syncom_colors2}
\alias{bugs.colors}
\alias{bugs.colors2}
\alias{bugs.colors3}
\title{Default Color Scheme}
\usage{
syncom_colors(x, v = NULL)

syncom_colors2(x, v = NULL)

bugs.colors(x, v = NULL)

bugs.colors2(x, v = NULL)

bugs.colors3(x, v = NULL)
}
\arguments{
\item{x}{Name of the variable type}

\item{v}{Optional. Vector of elements to color.}
}
\value{
Named character vector of default colors
}
\description{
Default colors for different variables.
}
\examples{
\dontrun{
col <- syncom_colors("OTU")
}
\dontrun{
col <- syncom_colors2("Species")
}
\dontrun{
col <- bugs.colors("Species")
}
\dontrun{
col <- bugs.colors2("Species")
}
\dontrun{
col <- bugs.colors2("Species")
}
}
\references{
See citation("syncomR")
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Theme}
\keyword{elements}
