% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{computeRL}
\alias{computeRL}
\title{Utility for stability calculations resilience}
\usage{
computeRL(dx0xt, dx0x1)
}
\arguments{
\item{dx0xt}{internal}

\item{dx0x1}{internal}
}
\description{
Function for computing resilience
}
\references{
\itemize{
\item{}{"Liu, Z., Cichocki, N., Bonk, F., Günther, S., Schattenberg, F., Harms, H., ... & Müller, S. (2018). Ecological stability properties of
microbial communities assessed by flow cytometry. mSphere, 3(1), e00564-17.
http://msphere.asm.org/content/3/1/e00564-17
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
Used internally by \code{\link{stability_properties}}
}
\keyword{Utilities}
