% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxa.R
\name{find_taxa}
\alias{find_taxa}
\title{Find Taxa That is Highest or Lowest Abundant in Each Sample}
\usage{
find_taxa(ps, which.taxa = c("most_abund", "least_abund"))
}
\arguments{
\item{ps}{a \code{\link{phyloseq-class}} with compositional data}

\item{which.taxa}{find "most_abund" or "least_abund"}
}
\value{
A data frame with sample_id, taxon highest or lowest abundance,
        and abundance value.
}
\description{
Identifying either taxon with highest or lowest abundance.
}
\details{
The output of this function can be used for identifying either
         taxon with highest or lowest abundance.
}
\examples{

data(SyncomFiltData)
ps.rel <- microbiome::transform(SyncomFiltData, "compositional")
otu.tb <- find_taxa(ps.rel, which.taxa = "most_abund")
head(otu.tb)
}
\references{
\itemize{
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Utilities}
