% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_members}
\alias{get_members}
\title{Utility for stability calculations}
\usage{
get_members(mergeTable, idx)
}
\arguments{
\item{mergeTable}{input from stability_properties}

\item{idx}{time id for stability_properties}
}
\value{
Taxa list and reference states as calculated by user provided distance metric from \code{\link{vegan}}.
}
\description{
List taxa for stability properties analysis as well as identifies reference states.
}
\details{
These are utility functions that are required for \code{\link{stability_properties}} analysis
}
\references{
\itemize{
\item{}{"Liu, Z., Cichocki, N., Bonk, F., Günther, S., Schattenberg, F., Harms, H., ... & Müller, S. (2018). Ecological stability properties of
microbial communities assessed by flow cytometry. mSphere, 3(1), e00564-17.
http://msphere.asm.org/content/3/1/e00564-17
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
To be used internally by \code{\link{stability_properties}}
}
\keyword{Utilities}
