% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_reference_phase_end}
\alias{get_reference_phase_end}
\title{Utility for stability calculations}
\usage{
get_reference_phase_end(data, begin, end)
}
\arguments{
\item{data}{internal stability_properties}

\item{begin}{internal stability_properties}

\item{end}{internal stability_properties}
}
\value{
Taxa list and reference states as calulated by user provided distance metric from \code{\link{vegan}}.
}
\description{
List taxa for stability properties analysis as well as identifies reference states.
}
\details{
Fundamental niche values are calculated on the basis of a multi-variate
}
\references{
\itemize{
\item{}{"Liu, Z., Cichocki, N., Bonk, F., Günther, S., Schattenberg, F., Harms, H., ... & Müller, S. (2018). Ecological stability properties of
microbial communities assessed by flow cytometry. mSphere, 3(1), e00564-17.
http://msphere.asm.org/content/3/1/e00564-17
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
Used internally by \code{\link{stability_properties}}
}
\keyword{Utilities}
