% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_variable.R
\name{group_by_variable}
\alias{group_by_variable}
\title{Utility for Formatting SynComGMM and SynComRNAKEGG Data}
\usage{
group_by_variable(x = NULL, by = c("KOID", "Module"))
}
\arguments{
\item{x}{A data.frame to format. Data frames similar to SynComRNAKEGG.}

\item{by}{Either KOID or Module}
}
\value{
A data frame.
}
\description{
Group by GMM ID or KO ID.
}
\details{
These are utility functions.
}
\references{
\itemize{
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
To be used internally
}
\keyword{Utilities}
