% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap.R
\name{niche_overlap}
\alias{niche_overlap}
\title{Niche Overlap Estimation Quantitative Traits from Temporal Abundance Data}
\usage{
niche_overlap(trab, graph = T, colorsN = NA)
}
\arguments{
\item{trab}{a matrix (x, N+1) with the abundances of the N species at x values of the
considered trait, NAs are not allowed in traits values.}

\item{graph}{TRUE or FALSE  indicating whether a graphical illustration of niche overlap is wanted.}

\item{colorsN}{vector (length=N) for colors of niches in hexadecimal code (if NA default colors).}
}
\value{
A plot or a data frame with the value of the overlap between corrected
        densities: 0 (no overlap) to 1(N identic densities).
}
\description{
Compute the niche overlap using quantitative traits Mouillot et al. 2005.
}
\details{
This approach was developed by Mouillot et al. 2005 and The code was
         obtained and modified from
         \url{https://github.com/umr-marbec/nicheoverlap/blob/master/nicheoverlap.R}.
         Computation is done between N species along a trait axis (functional or
         environmental) generalization of the Niche Overlap index (NO) based on kernel
         density estimation.  The index is non-parametric and assumes no normal distribution.
         This is particularly important as -omics data are known for challenges w.r.t normality.
         The N densities along the finite range of values considered are corrected according
         to their respective integrals, so that integrals of corrected densities along the range
         equal 1. If plot is T,each species niche have its own color, their overlap being shaded
         by lines.
}
\examples{
data(SyncomFiltData)
}
\references{
\itemize{
\item{}{Mouillot D, et al. Niche overlap estimates based on quantitative functional traits: a new
family of non-parametric indices. Oecologia. 2005 Sep 1;145(3):345-53.
\url{https://link.springer.com/article/10.1007/s00442-005-0151-z}}
\item{}{Shetty SA. et al. A Minimalist Approach for Deciphering the Ecophysiology of Human Gut Microbes (2020)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
