% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_seqtime.R
\name{normalize_seqtime}
\alias{normalize_seqtime}
\title{Normalize a matrix}
\usage{
normalize_seqtime(x, removeZero = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{removeZero}{Remove columns summing to zero.}
}
\value{
A normalized matrix
}
\description{
Normalize a matrix column-wise by dividing each entry by its corresponding column sum. Transfered
the function from seqtime R package \url{https://github.com/hallucigenia-sparsa/seqtime/blob/master/R/normalize.R}
becasue of issues with igraph dependency.
}
\details{
Columns summing to zero are removed by default.
}
\keyword{Data}
\keyword{and}
\keyword{filtering}
