% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module_abundances.R
\name{plot_module_abundances}
\alias{plot_module_abundances}
\title{Plot Temporal Changes in Module Abundances}
\usage{
plot_module_abundances(
  x,
  tax.variable = NULL,
  mm.variable = NULL,
  color.pal = NULL,
  nrow,
  ncol
)
}
\arguments{
\item{x}{A data.frame with Taxon, Module, Timepoint, Counts labels as value column}

\item{tax.variable}{Taxa to plot}

\item{mm.variable}{Module to plot}

\item{color.pal}{List of hex color codes}

\item{nrow}{When multiple values for modules and taxa}

\item{ncol}{When multiple values for modules and taxa}
}
\description{
Temporal taxa abundances are plotted for visual exploration.
}
\examples{
# data(SyncomGMM)
# focal.species <- c("Akkermansia_muciniphila",
#                    "Bacteroides_xylanisolvens",
#                    "Bacteroides_ovatus")

# p <- plot_module_abundances(SyncomGMM,
#  tax.variable = focal.species,
# mm.variable = c("propionate production III", "mucin degradation"),
# color.pal = strain.colors,
# nrow = 3,
# ncol = 1
# )
# p + theme_syncom() + theme(legend.position = "bottom")
}
\references{
\itemize{
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
