% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmds_stability.R
\name{plot_nmds_stability}
\alias{plot_nmds_stability}
\title{Plot Stability Properties NMDS}
\usage{
plot_nmds_stability(
  stab.in,
  refphase = "grey60",
  refstate = "red",
  evolution.points = "yellow",
  max.distant.states.euc = "steelblue",
  final.state = "black",
  max.distant.states.can = "brown3"
)
}
\arguments{
\item{stab.in}{The input come from \code{\link{stability_properties}} function.}

\item{refphase}{Color for reference phase. DEFAULT = grey60}

\item{refstate}{Color for reference state DEFAULT = red}

\item{evolution.points}{Color for to track evolution points DEFAULT = yellow}

\item{max.distant.states.euc}{Color for sites or timepoint that are max distant,
Euclidean DEFAULT = steelblue}

\item{final.state}{Color for last time state DEFAULT = black}

\item{max.distant.states.can}{Color for sites or timepoint that are max distant,
Canberra. DEFAULT = brown3}
}
\value{
A data frame and ggplot object.
}
\description{
NMDS plot compared to reference states as described by
             Florian Centler and Zishu Liu, UFZ Leipzig, Germany.
}
\details{
This script was modified to get ggplot object from the paper
         on "Ecological Stability Properties.Microbial Communities Assessed
         by Flow Cytometry" by Liu et al., 2018
         \url{http://msphere.asm.org/content/3/1/e00564-17}.
         This functions takes output from the \code{\link{stability_properties}}.

\itemize{
\item{}{Plot reference states and community evolusion}
}
}
\note{
\itemize{
\item{}{refphase= Color for reference phase. DEFAULT = grey60}
\item{}{refstate= Color for reference state DEFAULT = red}
\item{}{evolution.points= Color for to track evolution points DEFAULT = yellow}
\item{}{max.distant.states.euc= Color for sites/timepoint that are max distant,
        Euclidean DEFAULT = steelblue}
\item{}{max.distant.states.can = Color for sites/timepoint that are max distant,
        Canberra. DEFAULT = brown3}
\item{}{final.state= Color for last time state DEFAULT = black}
\item{}{To cite the package, see citation('syncomR')}
}
}
\examples{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
ps1.sub <- subset_samples(ps1.b5, Time_hr_num >= 172)
dat.stab <- stability_properties(ps1.sub, time.col = "Time_hr")
nmds.dt <- plot_nmds_stability(dat.stab)
print(nmds.dt$plot)
}
\references{
\itemize{
\item{}{Liu, Z., et al. (2018). Ecological stability properties of
        microbial communities assessed by flow cytometry. \emph{mSphere}, 3(1), e00564-17.
\url{http://msphere.asm.org/content/3/1/e00564-17}
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
