% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resilience_online.R
\name{plot_resilience_online}
\alias{plot_resilience_online}
\title{Plot Stability Properties Online Resilience}
\usage{
plot_resilience_online(stab.in, col.low = "#fa9fb5", col.high = "#49006a")
}
\arguments{
\item{stab.in}{The input come from \code{\link{stability_properties}} function.}

\item{col.low}{color aesthetics.}

\item{col.high}{color aesthetics.}
}
\value{
ggplot object.
}
\description{
Online resilience of the community as described by
             Florian Centler and Zishu Liu, UFZ Leipzig, Germany.
}
\details{
This script was modified to get ggplot object from the paper on
         "Ecological Stability Properties.Microbial Communities Assessed by
         Flow Cytometry" by Liu et al., 2018
         \url{http://msphere.asm.org/content/3/1/e00564-17}. This functions takes
         output from the \code{\link{stability_properties}}. The closer it is to zero
         the system is deviating continuously away from reference state, while values
         closer to 1 indicate the onset of recovery (1 = perfect recovery).
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
ps1.sub <- subset_samples(ps1.b5, Time_hr_num >= 172)
dat.stab <- stability_properties(ps1.sub, time.col = "Time_hr")
plot_resilience_online(dat.stab, col.low = "#fa9fb5", col.high = "#49006a")
}
}
\references{
\itemize{
\item{}{Liu, Z., et al. (2018). Ecological stability properties of
        microbial communities assessed by flow cytometry. \emph{mSphere}, 3(1), e00564-17.
\url{http://msphere.asm.org/content/3/1/e00564-17}
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
Input for this functions comes from \code{\link{stability_properties}}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
