% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_lineplot.R
\name{plot_species_lineplot}
\alias{plot_species_lineplot}
\title{Line Plot Species}
\usage{
plot_species_lineplot(
  ps,
  type = c("grouped", "individual"),
  group.var = NULL,
  time.col = "Time_hr",
  taxa.level = "OTU",
  ...
)
}
\arguments{
\item{ps}{A \code{\link{phyloseq-class}}}

\item{type}{Type of plot either area, line or bar}

\item{group.var}{Grouping variable like bioreactor}

\item{time.col}{Column specifying time variable}

\item{taxa.level}{Taxonomic level}

\item{...}{For facet_wrap}
}
\value{
ggplot object.
}
\description{
Temporal community composition using line plot.
}
\details{
Temporal community composition using line plot
}
\examples{
\dontrun{
data(SyncomFiltData)
SyncomFiltData.rel <- microbiome::transform(SyncomFiltData, "compositional")
}
}
\references{
\itemize{
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Anlaysis}
\keyword{and}
\keyword{visualization}
