% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stability_properties.R
\name{plot_stability_properties}
\alias{plot_stability_properties}
\title{Plot stability properties}
\usage{
plot_stability_properties(
  stab.in,
  property = c("dist.ot", "resilience.ot.eucl", "resilience.ot.canb",
    "resilience.oline")
)
}
\arguments{
\item{stab.in}{The input come from \code{\link{stability_properties}} function.}

\item{property}{c("dist.ot", "resilience.ot.eucl", "resilience.ot.canb", "resilience.oline")
for more details check original work by Liu et al., 2018 \url{\link{http://msphere.asm.org/content/3/1/e00564-17}}
\itemize{
\item{}{dist.ot= Plot dissimilarity over time}
\item{}{resilience.ot.eucl= Resilience using elucidean distance}
\item{}{resilience.ot.canb= Resilience using Canberra}
}}
}
\value{
Plots.
}
\description{
Plot various stability properties decribed by Liu et al., 2018.
}
\details{
This script was modified to from the paper on "Ecological Stability Properties.
Microbial Communities Assessed by Flow Cytometry" by Liu et al., 2018
 \url{\link{http://msphere.asm.org/content/3/1/e00564-17}}.
 This functions takes output from the \code{\link{stability_properties}}.
\itemize{
\item{}{Plot reference states and community evolution}
}
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
ps1.sub <- subset_samples(ps1.b5, Time_hr_num >= 28)
dat.stab <- stability_properties(ps1.sub,
  time.col = "Time_hr",
  experimentStart = 52, tref = 152
)
plot_stability_properties(dat.stab, property = "dist.ot")
}
}
\references{
\itemize{
\item{}{Liu, Z., et al. (2018). Ecological stability properties of
        microbial communities assessed by flow cytometry. \emph{mSphere}, 3(1), e00564-17.
\url{http://msphere.asm.org/content/3/1/e00564-17}
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
Input for this functions come from \code{\link{stability_properties}}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
