% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_syncom_composition.R
\name{plot_syncom_composition}
\alias{plot_syncom_composition}
\title{Plot temporal changes in taxa composition}
\usage{
plot_syncom_composition(
  ps,
  type = c("area", "bar"),
  time.col = "Time_hr",
  taxa.level = "OTU",
  sp.fill.pal = NULL,
  facet.var = NULL,
  ...
)
}
\arguments{
\item{ps}{A \code{\link{phyloseq-class}}}

\item{type}{Type of plot either area or bar}

\item{time.col}{Column specifying time variable}

\item{taxa.level}{Taxonomic level}

\item{sp.fill.pal}{List of colors for plotting}

\item{facet.var}{Column specifying varible to facet by. Here, its by Bioreactors}

\item{...}{options to pass for ggplot2 facet_}
}
\value{
ggplot object.
}
\description{
Temporal community composition.
}
\details{
Temporal community composition
}
\examples{
\dontrun{
data(SyncomFiltData)
SyncomFiltData.rel <- microbiome::transform(SyncomFiltData, "compositional")
strain.colors <- syncom_colors("BacterialSpecies")[1:16]
p <- plot_syncom_composition(SyncomFiltData.rel,
  type = "bar",
  time.col = "Time_hr_num",
  taxa.level = "OTU",
  sp.fill.pal = strain.colors, facet.var = "StudyIdentifier"
)
print(p)
}
}
\references{
\itemize{
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Anlaysis}
\keyword{and}
\keyword{visualization}
