% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot Temporal Changes in Taxa Abundances}
\usage{
plot_trajectory(
  ps,
  taxa.level = "Species",
  type = c("all", "single"),
  taxa = NULL,
  time.col = "Time_hr",
  group.variable = NULL,
  color.pal = NULL,
  transform.counts = "compositional"
)
}
\arguments{
\item{ps}{A \code{\link{phyloseq-class}}}

\item{taxa.level}{Taxonomic level}

\item{type}{Type of plot either all or single. If `type = single` taxa must be specified}

\item{taxa}{If single taxa name based on taxa.level that's choosen must be specified}

\item{time.col}{Column specifying time variable}

\item{group.variable}{Column specifying variable to be colored.
For e.g. if multiple subjects/bioreactors.}

\item{color.pal}{List of hex color codes}

\item{transform.counts}{Any one of \code{\link{transform}}}
}
\value{
ggplot object.
}
\description{
Temporal taxa abundances are plotted for visual exploration.
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.sycom.rel <- microbiome::transform(SyncomFiltData, "compositional")
fer_cols <- c(
  `Bioreactor A` = "#b2182b",
  `Bioreactor B` = "#2166ac",
  `Bioreactor C` = "#35978f"
)

tax.trac <- plot_trajectory(ps1.sycom.rel,
  time.col = "Time_hr",
  taxa.level = "Species",
  type = "single",
  taxa = "Akkermansia_muciniphila",
  group.variable = "StudyIdentifier",
  color.pal = fer_cols,
  transform.counts = "compositional"
)
tax.trac + geom_smooth() + theme_syncom()

}

}
\references{
\itemize{
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
