% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_properties.R
\name{stability_properties}
\alias{stability_properties}
\title{Community-level Stability Properties}
\usage{
stability_properties(
  ps,
  experimentStart = -1,
  experimentEnd = -1,
  overrideMaxEuclidean = -1,
  overrideMaxCanberra = -1,
  tref = -1,
  time.col = "Time_hr"
)
}
\arguments{
\item{ps}{A \code{\link{phyloseq-class}}. The input \code{\link{phyloseq}} object
must contain a single disturbance experiment (i.e. start and end of the
experiment refer to the first and last entry, respectively).}

\item{experimentStart}{Start of the experiment (experimentStart). A value of -1 will let
the script select these time points automatically. Otherwise,
specify a time value which is actually present in the data, i.e.
a value which appears in the first column of the input file.}

\item{experimentEnd}{End of the experiment (experimentEnd). A value of -1 will let
                     the script select these time points automatically. Otherwise,
                     specify a time value which is actually present in the data, i.e.
a value which appears in the first column of the input file.}

\item{overrideMaxEuclidean}{Maximal deviation (d_max) using Euclidean distance
(overrideMaxEuclidean). A value of -1 will let the script
select these time points automatically. Otherwise, specify
a time value which is actually present in the data, i.e.
a value which appears in the first column of the input file.}

\item{overrideMaxCanberra}{Maximal deviation (d_max) using Canberra distance
(overrideMaxCanberra). A value of -1 will let the script select
these time points automatically. Otherwise, specify a time value
which is actually present in the data, i.e. a value which appears
in the first column of the input file.}

\item{tref}{Disturbance event (tref). Otherwise, specify a time value which is actually present
in the data, i.e. a value which appears in the first column of the input file.}

\item{time.col}{Name of column in sample_data i.e. metadata specifying
time points.}
}
\value{
A data frame with stability properties.
}
\description{
Calculate community stability properties as decribed by
             Florian Centler and Zishu Liu, UFZ Leipzig, Germany.
}
\details{
This script was modified from the paper on "Ecological Stability Properties.
         Microbial Communities Assessed by Flow Cytometry" by Liu et al., 2018
         \url{http://msphere.asm.org/content/3/1/e00564-17}.
         Please cite this work when using this function. A value of -1 will let
         the script select these time points automatically,
         assuming that the file contains a single disturbance experiment
         (i.e. start and end of the experiment refer to the first and last entry,
         respectively). Otherwise, specify a time value which is actually present in
         the data, i.e.a value which appears in the first column of the input file.

\itemize{
\item{}{Identify reference states}
\item{}{To cite the package, see citation('syncomR')}
}
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
ps1.sub <- subset_samples(ps1.b5, Time_hr_num >= 120)
dat.stab <- stability_properties(ps1.sub, time.col = "Time_hr")
head(dat.stab)
}
}
\references{
\itemize{
\item{}{Liu, Z., et al. (2018). Ecological stability properties of
        microbial communities assessed by flow cytometry. \emph{mSphere}, 3(1), e00564-17.
\url{http://msphere.asm.org/content/3/1/e00564-17}
}
\item{}{To cite the package, see citation('syncomR')}
}
}
\seealso{
To be used internally by \code{\link{stability_properties}}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Anlaysis}
\keyword{and}
\keyword{visualization}
