% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{syncom_colors}
\alias{syncom_colors}
\title{Default Color Scheme}
\usage{
syncom_colors(x, v = NULL)
}
\arguments{
\item{x}{Name of the variable type}

\item{v}{Optional. Vector of elements to color.}
}
\value{
Named character vector of default colors
}
\description{
Default colors for different variables.
}
\examples{
\dontrun{
col <- syncom_colors("OTU")
}
}
\references{
See citation("syncomR")
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Theme}
\keyword{elements}
