% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SyncomCopyCorrectedCounts}
\alias{SyncomCopyCorrectedCounts}
\title{SynComm Copy Number Corrected Species-level data}
\format{
List of the following data matrices as described in detail in
Shetty SA et al. (20XX):
\itemize{
\item HPLC: Quantification of metabolites majorly short chain fatty acids
\item microbes: 16S rRNA gene counts corrected for copy numbers of
      16 species-level counts across 210 samples
\item meta: Sample metadata including time point, StudyIdentifier, Perturbations, etc
\item phyloseq The syn comm data set converted into a
\code{\link{phyloseq-class}} object.
}
}
\usage{
data(SyncomCopyCorrectedCounts)
}
\value{
Loads the data set in R.

Loads the data set in R.
}
\description{
The \strong{SyncomCopyCorrectedCounts} data set contains high-through
             taxonomic profiling data from 186 samples taken from the syncom
             experiment and 16 species-level counts corrected with qPCR data.
             For raw ASV-level counts check \strong{SyncomRawCounts}.
}
\references{
Shetty SA et al. (20XX) XXX 1:e32
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{data}
