% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SyncomGMM}
\alias{SyncomGMM}
\title{SyncomGMM Gut Metabolic Modules (GMMs)}
\format{
List of the following data matrices as described in detail in
       Shetty SA et al. (20XX):
}
\usage{
data(SyncomGMM)
}
\value{
Loads the data set in R.

Loads the data set in R.
}
\description{
The \strong{SyncomGMM} consists 36 samples and abundances of GMMs
linked with taxa. This was obtained by running the \emph{omixerRpm} on
KEGG KO abundances and stored as data.frame object and made available in
\emph{syncomR} package.
}
\references{
Shetty SA et al. (20XX) XXX 1:e32
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{data}
