% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SyncomRawCounts}
\alias{SyncomRawCounts}
\title{SynComm Raw ASV Data}
\format{
List of the following data matrices as described in detail in
Shetty SA et al. (20XX):
\itemize{
\item HPLC: Quantification of metabolites majorly short chain fatty acids
\item microbes: 16S rRNA gene counts of 5269 ASV-level taxa across 210 samples
\item meta: Sample metadata including time point, StudyIdentifier, Perturbations, etc
\item phyloseq The syn comm data set converted into a
\code{\link{phyloseq-class}} object.
}
}
\usage{
data(SyncomRawCounts)
}
\value{
Loads the data set in R.
}
\description{
The \strong{SyncomRawCounts} data set contains high-through taxonomic profiling data from
             210 samples taken from the syncom experiment and 5269 ASV-level counts.
             The data set includes control samples samples from feed,
             reagent and DNA extraction procedure.
             For details, see reference.
}
\references{
Shetty SA et al. (20XX) XXX 1:e32
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
