% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SynComRawRNA}
\alias{SynComRawRNA}
\title{SynComm Raw Locus Tag Count From Raw Read Processing}
\format{
List of the following data matrices as described in detail in
Shetty SA et al. (20XX):
}
\usage{
data(SynComRawRNA)
}
\value{
Loads the data set in R.

Loads the data set in R.
}
\description{
The \strong{SynComRawRNA} consist of raw counts from RNAseq
             for 36 samples (51138 locus tags) stored as data.frame
             object and made available in \emph{syncomR} package.
}
\references{
Shetty SA et al. (20XX) XXX 1:e32
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
