% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SynComRNAKEGG}
\alias{SynComRNAKEGG}
\title{SynComm KEGG lLinking of SynComRawRNA}
\format{
List of the following data matrices as described in detail in
        Shetty SA et al. (20XX):
}
\usage{
data(SynComRNAKEGG)
}
\value{
Loads the data set in R.

Loads the data set in R.
}
\description{
The \strong{SynComRNAKEGG} consists 36 samples and locus tags
             linked to KEGG IDs with following columns KO, LocusTag,
             BacterialStrain, GeneName, Level_1, Level_2, Level_3, KOID,
             KOGeneName stored as data.frame object and made available
             in \emph{syncomR}.
}
\references{
Shetty SA et al. (20XX) XXX 1:e32
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{data}
