% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_time_table.R
\name{taxa_time_table}
\alias{taxa_time_table}
\title{Make taxa abundance table}
\usage{
taxa_time_table(
  ps,
  normalize = c(TRUE, FALSE),
  time.col = NULL,
  remove.zero = c(TRUE, FALSE)
)
}
\arguments{
\item{ps}{a \code{\link{phyloseq-class}}}

\item{normalize}{TRUE or FALSE uses the normalize function from \emph{seqtime} by Faust et al.}

\item{time.col}{Specify column containing time variable}

\item{remove.zero}{TRUE or FALSE}
}
\value{
Normalized or raw counts taxa abundance table (taxaare rows and timepoits columns).
}
\description{
Converts \code{\link{phyloseq-class}} object to taxa table arranged by time point.
}
\details{
The output of this function can be used for functions that require taxa tables in matrix form.
}
\examples{
\dontrun{
data(SyncomFiltData)
ps1.b5 <- subset_samples(SyncomFiltData, StudyIdentifier == "Bioreactor A")
remove_T80 <- c("Ferm_1_5_80", "Ferm_1_6_80", "Ferm_1_8_80")
ps1.sub <- prune_samples(!(sample_names(ps1.b5) \%in\% remove_T80), ps1.b5)
otu.tb <- taxa_time_table(ps1.sub, normalize = TRUE, time.col = "Time_hr", remove.zero = TRUE)
head(otu.tb)
}
}
\references{
\itemize{
\item{}{Faust, K., et al. (2018). Signatures of ecological processes in
microbial community time series. Microbiome, 6(1), 120.
\url{https://doi.org/10.1186/s40168-018-0496-2}}
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
