% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_diversity.R
\name{temporal_diversity}
\alias{temporal_diversity}
\title{Measure and plot temporal changes in eveness of community}
\usage{
temporal_diversity(
  ps,
  time.col = "Time_hr_num",
  div.measure = "gini",
  plot = TRUE
)
}
\arguments{
\item{ps}{a otu/ASV/species abundance table}

\item{time.col}{Column specifying time variable}

\item{div.measure}{Currently only Gini as a measure of inequality is used}

\item{plot}{TRUE or FALSE if TRUE plot is returned else a data frame}
}
\value{
Temporal plot of inequality.
}
\description{
Calculate and plot diversity measures over time.
}
\details{
Can be useful to identify potential trends in temporal community diversities.
}
\examples{
\dontrun{
data(SyncomFiltData)
p <- temporal_diversity(SyncomFiltData, time.col = "Time_hr_num", div.measure = "gini")
p + facet_grid(~StudyIdentifier) + geom_smooth(fill = "#a6bddb", color = "#d94801")
}

}
\references{
\itemize{
\item{}{To cite the package, see citation('microbiome')}
\item{}{'Shetty SA et al (2019-2024)}
\item{}{To cite the package, see citation('syncomR')}
}
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Analysis}
\keyword{and}
\keyword{visualization}
