% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{theme_syncom}
\alias{theme_syncom}
\title{Default ggplot2 theme}
\usage{
theme_syncom(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_size}{Similiar to ggplot2 themes}

\item{base_family}{Similiar to ggplot2 themes}

\item{base_line_size}{Similiar to ggplot2 themes}

\item{base_rect_size}{Similiar to ggplot2 themes}
}
\description{
Default ggplot2 theme for consistency in publications.
}
\references{
See citation("syncomR")
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Theme}
\keyword{elements}
